Ciwan

+++ App/app/src/main/java/com/projectc/mythicalmonstermatch/CardClass.java
@@ -13,10 +13,10 @@ import java.util.Map;
 public class CardClass{
     public int id;                                                                             //ID zwischen 1-30 zur leichten eindeutigen identifiezierung der Karte
     public String name;                                                                        //Name des Monster (z.B. Medusa, Golem etc)
-    public Map<String, Float> attributeMap = new HashMap<String, Float>();                     //Attribute Map zum schlanken Vergleich Werte da über String Callbar
+    public Map<String, Integer> attributeMap = new HashMap<String, Integer>();                     //Attribute Map zum schlanken Vergleich Werte da über String Callbar
     public Bitmap img;                                                                          //Anzeige Bild, evtl auslagern TODO Nachforschen Möglichkeiten anzeigen
 
-    public CardClass(int id, String name, float a1, float a2, float a3, float a4, float a5, Bitmap img){
+    public CardClass(int id, String name, int a1, int a2, int a3, int a4, int a5, Bitmap img){
         this.id = id;
         this.name = name;
         this.attributeMap.put("attribute1", a1);                                    //Platzhalterattributesnamen da noch nicht ausgewählt welche genommen werden
